Fix byte/unicode issues in smoketest/loggers.py
Writing a bytestring to sys.stdout throws a TypeError in Python3. This
resolves that problem by always writing unicode strings.

As I understand it, the encoding in this line:

    sys.stdout.write(full_message.encode('utf-8'))

was intended to address an issue where logging full response bodies
(verbosity >= 4) to files, either with the -o flag or shell redirection,
would cause a UnicodeEncodeError.

The -o flag should continue to work, but with this change, shell
redirection in Python2 may require setting PYTHONIOENCODING :(